requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
Werkzeug==2.3.7
Flask-Mail==0.10.0
dotenv==0.9.9
python-dotenv==1.1.1
psycopg2-binary>=2.9.0